tests/wpt/web-platform-tests/css/css-exclusions/wrap-flow-003.html

<!DOCTYPE HTML>
<html>
    <head>
        <title>CSS Exclusions Test: wrap-flow property set to auto</title>
        <link rel="author" title="Jacob Goldstein" href="mailto:jacobg@adobe.com">
        <link rel="help" href="http://www.w3.org/TR/css3-exclusions/#wrap-flow-property">
        <meta name="flags" content="ahem dom">
        <meta name="assert" content="No exclusion should be created">
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="resources/helper.js"></script>
        <style>
            * {
                margin: 0px;
                padding: 0px;
            }
            .exclusion {
                wrap-flow: auto;
                position: absolute;
                top: 10px;
                left: 20px;
                width: 60px;
                height: 25px;
                background: blue;
            }
            #content {
                width: 400px;
                line-height: 12px;
                font: 12px Ahem;
            }
        </style>
    </head>
    <body>
        <div class="exclusion"></div>

        <div id="content">
            <span id="line1">ethical beer craft</span><BR>
            <span id="line2">Fixie pork belly</span><BR>
            <span id="line3">food truck in marfa</span><BR>
            <span id="line4">american apparel squid</span>
        </div>

        <div id="log"></div>

        <script type="text/javascript">
            test(function() {assert_equals(checkLinePos("line1",216,"right"), true)}, "Verify right of the 'line1' span is positioned correctly");
            test(function() {assert_equals(checkLinePos("line2",192,"right"), true)}, "Verify right of the 'line2' span is positioned correctly");
            test(function() {assert_equals(checkLinePos("line3",228,"right"), true)}, "Verify right of the 'line3' span is positioned correctly");
            test(function() {assert_equals(checkLinePos("line4",36,"top"), true)}, "Verify top of the 'line4' span is positioned correctly");
            test(function() {assert_equals(checkLinePos("line4",264,"right"), true)}, "Verify right of the 'line4' span is positioned correctly");
        </script>
    </body>
</html>
